ConsulNamingService with acl change base version to 3.0.3-SNAPSHOT
