+++ brpc-java-examples/brpc-java-core-examples/src/main/java/com/baidu/brpc/example/standard/SyncBenchmarkTest.java
@@ -46,8 +46,6 @@ public class SyncBenchmarkTest {
     }
 
     public static void main(String[] args) throws InterruptedException, IOException {
-        args=new String[]{"list://127.0.0.1:8002","2"};
-
         if (args.length != 2) {
             System.out.println("usage: BenchmarkTest list://127.0.0.1:8002 threadNum");
             System.exit(-1);
@@ -67,7 +65,7 @@ public class SyncBenchmarkTest {
         int threadNum = Integer.parseInt(args[1]);
 
         InputStream inputStream = Thread.currentThread().getClass()
-                .getResourceAsStream("/message_4k.txt");
+                .getResourceAsStream("/message.txt");
         int length = inputStream.available();
         byte[] messageBytes = new byte[length];
         inputStream.read(messageBytes);
